Neovim and fish Done with safeguards

+++ neovimCustom/deleteNeo.sh
@@ -1,3 +1,15 @@
 #!/bin/bash
-sudo rm -rf ~/.config/nvim
-sudo apt purge neovim
+if [  -d "$HOME/.config/nvim" ]; then
+	rm -rf ~/.config/nvim
+	echo "neovim config removed"
+else
+	echo "neovim config doesn't exist"
+fi
+
+if [ -f "/usr/bin/nvim" ]; then
+	sudo apt purge neovim
+	sudo apt autoremove
+	echo "neovim removed"
+else
+	echo "neovim not installed"
+fi
